Refactor baserouter and create crudrouter
