--- src/api/roles.py
@@ -1,13 +1,0 @@
-from .baserouter import BaseRouter
-from src.models import Role, RolePydantic
-from src.models.schema.department import RoleSchema
-
-
-router = BaseRouter(
-    model=Role,
-    request_schema=RoleSchema,
-    response_schema=RolePydantic
-)
-
-router.load_crud_routes()
-
