+++ src/api/users.py
@@ -1,0 +1,7 @@
+from .baserouter import BaseRouter
+from src.models import User, UserPydantic
+from src.models.schema.user import UserSchema
+
+
+router.load_crud_routes()
+
