+++ src/routes.py
@@ -1,5 +1,5 @@
 from fastapi import Depends
-from src.api import auth, roles
+from src.api import auth, users
 from src.utils.security import get_current_user
 
 
@@ -18,7 +18,7 @@ def add_routers(app, config):
     )
 
     app.include_router(
-        roles.router,
+        users.router,
         prefix=f"{config.API_URL}/roles",
         tags=["Roles"],
         dependencies=[Depends(get_current_user)]
